backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
google-genai>=0.3.0
cloudinary